# Projekt/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Tasks",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "description": "Gets all tasks",
        "responses": {
          "200": {
            "description": "OK - returns all tasks"
          },
          "400": {
            "description": "Bad Request - when there are no tasks"
          },
          "401": {
            "description": "Unauthorized - if jwt token is invalid or not present"
          }
        }
      },
      "post": {
        "description": "Create new task",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "Homework",
                  "required": true
                },
                "description": {
                  "example": "Do Homework",
                  "required": true
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity - if either title or description are missing as both are required"
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "description": "Gets single task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity - if id does not exist"
          }
        }
      },
      "patch": {
        "description": "Update a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "Math Homework"
                },
                "description": {
                  "example": "Do Math Homework"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity - if both title and description are missing, at least something needs to be present to update"
          }
        }
      },
      "delete": {
        "description": "Delete a task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity - if id does not exist"
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "Login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "example": "any@mail.example"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity - if either user or password are missing"
          }
        }
      }
    },
    "/verify": {
      "get": {
        "description": "Verify token and see time until expiry",
        "responses": {
          "200": {
            "description": "OK - returns for how long the token is still valid"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/logout": {
      "delete": {
        "description": "Destroys the token",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}